Создал класс MessageService! Изменил метод setMessage в классе MessageDao!

+++ src/main/java/by/it_academy/jd2/Mk_JD2_95_22/servise/api/IMessageService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface IMessageService {
     boolean send(MessageDTO messageDTO);
     List<MessageDTO> getMessageForUser(String login);
-    int countUsers();
+    long countUsers();
 }
